reader : corrections de problèmes d'offsets
git-svn-id: http://rdinnovation.onf.fr/svn/computree@266 0d0f61bb-bbfd-4940-b5ed-fa75979e1372

+++ pluginshared/ct_reader/ct_reader_ascrgb.h
@@ -15,12 +15,12 @@ public:
     CT_Reader_ASCRGB();
 
     bool setFilePath(const QString &filepath);
-    void setRadiusFilter(const float &radius);
+    void setRadiusFilter(const double &radius);
 
     CT_AbstractReader* copy() const;
 
 private:
-    float   m_filterRadius;
+    double   m_filterRadius;
 
 protected:
     void protectedInit();
